requirements.txt
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
chromadb
pypdf
streamlit
python-dotenv

